% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Unit Hydrograph initializing}
\usage{
init(Timeresinsec, modelLayer, modelRiver, modelMAD)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{modelLayer}{list of parameters about the Layers
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e), NoL=f)}

\item{modelRiver}{list of parameters about the river
list(maxL=a,speed=b,nbStepsDelay=c ,z=d,distr="dnorm",param=c(e,f))}

\item{modelMAD}{list of parameters about the Mean Annual Discharge
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e))}
}
\description{
The funtion initialize the main features of the Unit Hydrograph:
- The UH of the river,
- the UH of the layers
- The UH of the Mean Annual Discharge (MAD)
}
\examples{
\dontrun{
init()
}
}
\keyword{UH}

