% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.UH.R
\name{init.UH}
\alias{init.UH}
\title{Unit Hydrograph initializing}
\usage{
init.UH(method = NULL, path = NULL, Timeresinsec, modelLayer, modelRiver,
  modelMAD)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "processed"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param, NoL)}

\item{modelRiver}{list of parameters about the river
list(maxL,speed,nbStepsDelay,z,distr,param)}

\item{modelMAD}{list of parameters about the Mean Annual Discharge
list(maxL,speed,nbStepsDelay,z,distr,param)}
}
\description{
The funtion initialize the main Unit Hydrographs:
- The UH of the river,
- the UH of the layers
- The UH of the Mean Annual Discharge (MAD)
}
\examples{
\dontrun{
init.UH()
}
}
\keyword{UH}

