% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerUH.R
\name{layerUH}
\alias{layerUH}
\title{Unit Hydrograph of Layers}
\usage{
layerUH(maxL, speed, Timeresinsec, model)
}
\arguments{
\item{maxL}{Max distance (from distance distribution, DD)}

\item{speed}{celerity}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{model}{list of parameters about the UH
example: \code{list(z=0.5,distr="dexp",param=c(midL))}
with:
- z: arealfraction of DD (what area with distance zero to river), midL: Mean distance (from distance distribution, DD) for soils
- distr: dexp, dnorm,...
- param: c(valpar1,valpar2)}
}
\description{
The function \code{layerUH()} compute the UH of the saturation layers
}
\examples{
\dontrun{
layerUH()
}
}
\keyword{UH}

