% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHslopes.R
\name{UHslopes}
\alias{UHslopes}
\title{Unit Hydrograph of slopes}
\usage{
UHslopes(layerUH, ddist)
}
\arguments{
\item{layerUH}{UH of the saturation layers}

\item{ddist}{states of each saturation level.}
}
\description{
The function \code{UHslopes()} compute the UH of the slopes
}
\examples{
\dontrun{
UHslopes()
}
}
\keyword{UH}

