% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHvec.R
\name{UHvec}
\alias{UHvec}
\title{Unit Hydrograph of x}
\usage{
UHvec(maxL, speed, Timeresinsec, model)
}
\arguments{
\item{maxL}{Max distance (from distance distribution, DD)}

\item{speed}{celerity}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{model}{model of the UH. \code{list(z,distr,param=c())}}
}
\description{
The function \code{UHvec()} computes a Unit Hydrograph
}
\examples{
\dontrun{
UHvec()
}
}
\keyword{UH}

