% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xConvRiver.R
\name{xConvRiver}
\alias{xConvRiver}
\title{Convolve UH with the one of the river}
\usage{
xConvRiver(UHx, UHriver)
}
\arguments{
\item{UHx}{Unit Hydrograph of a specific type of terrain}

\item{UHriver}{Unit Hydrograp of the river}
}
\description{
The function \code{xConvRiver()} convolves one UH with the one of the river
}
\examples{
\dontrun{
UH.xConvRiver()
}
}
\keyword{UH}

